Pass YAML contents to Yaml::parse
Make the call compatible with future Symfony version, and avoid
E_USER_DEPRECATED as thrown by the current 2.7.0-beta2:

The ability to pass file names to the Symfony\Component\Yaml\Yaml::parse
method is deprecated since version 2.2 and will be removed in 3.0. Pass
the YAML contents of the file instead.